// backend/sdin-hnbra/src/modules/users/dto/create-user-response.dto.ts
export class CreateUserDtoResponse {
    nip: string;
    patent: string;
    warName?: string;
    firstName: string;
    lastName: string;
    role: string;
    status: string;
    permission: string;
    password: string;
    emailPersonal?: string;
    emailMb?: string;
    contactNumber?: string;
}